requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-storage-blob==12.24.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-enum-choices==2.1.4
django-enumfields==2.1.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
drf-enum-field==0.9.3
idna==3.10
isodate==0.7.2
psycopg2-binary==2.9.10
pycparser==2.22
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
gunicorn==20.1.0
whitenoise==6.9.0